Created CounterFragment.updateEnemyCounterImgViews(), replacing redundant methods within Weapon & Char Fragments
